==> 01-Analyse des données.ipynb <==
import pandas as pd
df_initial = pd.read_csv('feux_de_foret.csv')
df_initial.head()

df_initial.info()

df_initial.duplicated().sum()


df_initial.isna().sum()

#De nombreuses variables nous seront a priori inutiles pour l'analyse, la visualisation ou l'élaboration d'un modèle de prédiction. Nous les supprimerons afin d'alléger le dataset.
#La plupart des variables utiles sont complètes. Toutefois, certaines coordonnées géographiques précises (le nom et le code FIPS du comté) et la date de contrôle du feu sont souvent manquantes.


#Cette fonction permet de déterminer le nombre de NAs par variable et par Etat. Il s'agit de vérifier si les NAs sont bien répartis.

STATE_LIST = sorted(list(df_initial.STATE.unique()))
df_NAs = pd.DataFrame()
df_NAs['STATES'] = STATE_LIST

def NA_Counters(colonne):
    
    col1 = str(colonne)+'.total'
    col2 = str(colonne)+'.percent'
    liste_total = []
    liste_percent = []
    
    for state in STATE_LIST:
       
        x = df_initial[df_initial['STATE'] == state][colonne].isna().sum()
        y = round((x/1880465)*100,2)
        liste_total.append(x)
        liste_percent.append(y)
        
    df_NAs[col1] = liste_total
    df_NAs[col2] = liste_percent
        
        
        

NA_Counters('DISCOVERY_TIME') 
NA_Counters('CONT_TIME') 
NA_Counters('FIPS_CODE') 
NA_Counters('FIPS_NAME') 

df_NAs.head()

#On remarque par exemple que la Californie a des taux de NAs élevés, alors qu'il s'agit de l'un des Etats les plus touchés par les feux.
#Nous stockons ces résultats dans un DataFrame afin de pouvoir les comparer plus facilement.

df_NAs.sort_values(by='FIPS_CODE.percent', axis=0, ascending=False).head()
