==> 08- Ajout des données météo.ipynb <==
import pandas as pd

df_PCPN = pd.read_csv('PCPN.csv')
df_TMAX = pd.read_csv('TMAX.csv')
df_TAVG = pd.read_csv('TAVG.csv')
df_TMAX = df_TMAX.rename(columns={'Unnamed: 0': 'ID', 'Unnamed: 1': 'Janvier', 'Unnamed: 2': 'Février', 'Unnamed: 3': 'Mars', 'Unnamed: 4': 'Avril', 'Unnamed: 5': 'Mai', 'Unnamed: 6': 'Juin', 'Unnamed: 7': 'Juillet', 'Unnamed: 8': 'Août', 'Unnamed: 9': 'Septembre', 'Unnamed: 10': 'Octobre', 'Unnamed: 11': 'Novembre', 'Unnamed: 12': 'Décembre'})

#Exemple de DataFrame récupéré :
df_PCPN.head()

#Création des DataFrames fusionnables:

def cleaning_df(df,string):
    
    #Dans un premier temps, nous scindons l'identifiant à 10 chiffres de chaque ligne.
    #Nous obtenons l'année, le mois, l'Etat et le comté de chaque mesure.
    
    df['ID'] = df.ID.astype(str).str.zfill(11)
    
    df['Etat'] = df.ID.astype(str).apply(lambda x:x[:2])
    
    df['Comté'] = df.ID.astype(str).apply(lambda x:x[2:5])

    df['Année'] = df.ID.astype(str).apply(lambda x:x[-4:])
    
    #Nous n'avons plus besoin de la colonne ID. 
    #Pour fusionner les trois dataframes, nous utiliserons les autres colonnes.
    df = df.drop('ID', axis = 1)
       
    
    Liste_mois = ['Janvier','Février','Mars','Avril','Mai','Juin','Juillet','Août','Septembre','Octobre','Novembre','Décembre']
    df = pd.melt(df,['Comté','Année','Etat'],Liste_mois)

    df = df.rename(columns={'value':string,'variable' : 'Mois'})
    
    df = df.loc[(df['Année'].astype(int) > 1991) & (df['Année'].astype(int) < 2016)]
    
    return df


df_PCPN_clean = cleaning_df(df_PCPN,'PCPN')
df_TMAX_clean = cleaning_df(df_TMAX,'TMAX')
df_TAVG_clean = cleaning_df(df_TAVG,'TAVG')




#Exemple de DataFrame transformé : 
df_TMAX_clean.head()

#Fusion des trois DataFrames en deux temps :

df_deux = pd.merge(left =df_TMAX_clean, right=df_TAVG_clean, how="inner", left_on=['Comté','Année','Etat','Mois'], right_on=['Comté','Année','Etat','Mois'])
df_ALL = pd.merge(left =df_deux, right=df_PCPN_clean, how="inner", left_on=['Comté','Année','Etat','Mois'], right_on=['Comté','Année','Etat','Mois'])
df_ALL.Année = df_ALL.Année.astype('int64')

#Création d'un DataFrame exportable:
df_ALL.to_csv('df_TMAX_TAVG_PCPN')


df_init = pd.read_csv('df_initial.csv',usecols=['FIRE_YEAR','DISCOVERY_DATE','FIPS_CODE','STATE','FIRE_SIZE','COUNTY'])

def julian_to_gregorian(julian_date):
    return pd.to_datetime(julian_date, origin='julian', unit='D').strftime('%d/%m/%Y')
# Appliquer la conversion à la colonne DISCOVERY_DATE
df_init['DISCOVERY_DATE'] = df_init['DISCOVERY_DATE'].apply(julian_to_gregorian)

#Cette manoeuvre provoque de nombreux avertissements qui ne sont pas importants pour la suite du travail.
import warnings
warnings.filterwarnings('ignore')
df_init.DISCOVERY_DATE = pd.to_datetime(df_init.DISCOVERY_DATE).dt.month

df_init = df_init.rename(columns={"DISCOVERY_DATE":"Mois","FIRE_YEAR":"Année",'STATE':'Etat','COUNTY':'Comté'})



Il faut alors doter chaque comté d'un code propre, constitué de son état et de son code FIPS.
Cette manoeuvre est nécessaire car les comtés peuvent partager un même nom d'un Etat à l'autre. 

#Création des listes en vue du remplacement. 
Liste_ID = ['01','02','03','04','05','06','07','08','09', 
            '10', '11','12','13','14','15','16','17','18','19', '20',
            '21','22','23','24','25','26','27','28','29', '30',
            '31','32','33','34','35','36','37','38','39', '40',
            '41','42','43','44','45','46','47','48','50']

Liste_Etats = ['AL','AZ','AR','CA','CO','CT','DE','FL','GA','ID','IL','IN','IA','KS','KY','LA','ME','MD','MA','MI','MN','MS','MO',
              'MT','NE','NV','NH','NJ','NM','NY','NC','ND','OH','OK','OR','PA','RI','SC','SD','TN','TX','UT','VT','VA','WA','WV','WI','WY','AK']


df_init.FIPS_CODE = df_init.FIPS_CODE.fillna(999)
df_init.Etat = df_init.Etat.replace(Liste_Etats,Liste_ID)

df_init.Comté = df_init.Etat.astype(str) + '-' + round(df_init.FIPS_CODE,0).astype(int).astype(str).str.zfill(3)

df_ALL.Mois = df_ALL.Mois.replace(['Janvier', 'Février', 'Mars', 'Avril', 'Mai', 'Juin', 'Juillet',
       'Août', 'Septembre', 'Octobre', 'Novembre', 'Décembre'],[1,2,3,4,5,6,7,8,9,10,11,12])
df_ALL['Etat_Comté'] = df_ALL.Etat.astype(str) + '-' + df_ALL.Comté.astype(str)
df_ALL.head()


Nous pouvons maintenant fusionner le dataframe créé et le dataframe original.

df_init.head()

df_init_groupé = df_init.groupby(['Année','Mois','Comté'],as_index=False).agg({'FIRE_SIZE':'sum','FIPS_CODE':'count'})
df_init_groupé.head()

df_FINAL = pd.merge(left = df_ALL, right = df_init_groupé, how = 'left', left_on = ['Année','Mois','Etat_Comté'], right_on = ['Année','Mois','Comté'])

df_FINAL = df_FINAL.drop(['Comté_x','Comté_y'],axis=1)
df_FINAL.FIPS_CODE = df_FINAL.FIPS_CODE.fillna(0).astype(int)
df_FINAL.FIRE_SIZE = df_FINAL.FIRE_SIZE.fillna(0)
df_FINAL['Comté'] = df_FINAL.Etat_Comté.apply(lambda x:x[3:])
df_FINAL = df_FINAL.rename(columns={'FIPS_CODE':'Nombre de feux','FIRE_SIZE':'Etendue totale des feux'})
df_FINAL.head()

df_FINAL.to_csv('Dataframe_Final_Comtés_Mois.csv')
