==> 11- Métriques de la prédiction du nombre de feux.ipynb <==
NOMBRE DE FEUX



#Score de chaque état avec RandomForestRegressor

from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import r2_score
from sklearn.model_selection import train_test_split
import pandas as pd
import joblib
df = pd.read_csv('df_trans.csv')

#Metriques
import sklearn.metrics
from sklearn.metrics import mean_squared_error
from sklearn.metrics import mean_absolute_error

#Getdummies
county_dum = pd.get_dummies(df['Comté'],prefix='County', prefix_sep='_')
#date_dum = pd.get_dummies(df['Année'],prefix='Y', prefix_sep='_')
mois_dum = pd.get_dummies(df['Mois'],prefix='M', prefix_sep='_')
df = df.join(mois_dum)
#df = df.join(date_dum)
df = df.join(county_dum)

liste = list(df.Etat.unique())

RRF = RandomForestRegressor(random_state=42) 

#Création DF
df_score = pd.DataFrame()
nom = []
Train_score = []
Test_score = []
MAE = []
MSE = []
R2 = []
Median = []
Quantile75 = []
Min = []
Max = []


 
liste = list(df.Etat.unique())


for state in liste:
    #Création du df
    name = str(state)
    data = df[df['Etat'] == state]
    
    #Séparation des données
    feats = data.drop(['Etendue totale des feux','Nombre de feux','Etat_Comté','Etat','Unnamed: 0', 'Mois','Comté'], axis = 1)
    target = data['Nombre de feux']
    
    feats_train, feats_test, target_train, target_test = train_test_split(feats, target, test_size = 0.2)
    
    #Entraînement du modèle    
    RRF.fit(feats_train, target_train)
    
    #Prédictions    
    target_pred_test = RRF.predict(feats_test)
    
    #Score
    score_train = RRF.score(feats_train,target_train)
    score_test = RRF.score(feats_test,target_test)
    mae_ = mean_absolute_error(target_test, target_pred_test)
    mse_ = mean_squared_error(target_test, target_pred_test,squared=True)
    R2_ = r2_score(target_test, target_pred_test)
    
    #Métriques du df
    médiane = data['Nombre de feux'].median()
    quantile = data['Nombre de feux'].quantile(q=0.75)
    minimum = data['Nombre de feux'].min()
    maximum = data['Nombre de feux'].max()
    
     
    #Ajout au df
    nom.append(name)
    Train_score.append(score_train)
    Test_score.append(score_test)
    MAE.append(mae_)
    MSE.append(mse_)
    R2.append(R2_)
    
    Median.append(médiane)
    Quantile75.append(quantile)
    Min.append(minimum)
    Max.append(maximum)
    
    #joblib
    joblib.dump(RRF, f'{name}'+'model.joblib')
    
    
#Création df_score    

df_score['Etat'] = nom
df_score['Score Train'] = Train_score
df_score['Score Test'] = Test_score
df_score['MAE'] = MAE
df_score['MSE'] = MSE
df_score['R2'] = R2
df_score['Médiane'] = Median
df_score['3è quartile'] = Quantile75
df_score['Minimum'] = Min
df_score['Maximum'] = Max



df_score.sort_values(by='Score Test', ascending = False)
    
print(f"Moyenne du score de test : {round(df_score['Score Test'].mean(),2)}, Moyenne de la MAE : {round(df_score['MAE'].mean(),2)}, Moyenne de la MSE : {round(df_score['MSE'].mean(),2)}") 
df_score
